apps/auditor/requirements.txt
flask
requests
psycopg2-binary
redis